+++ eon_timer/timers/enhanced_entralink_timer.py
@@ -15,9 +15,9 @@ class EnhancedEntralinkTimer:
                target_delay: int,
                target_second: int,
                target_advances: int,
-               calibration: float,
-               entralink_calibration: float,
-               frame_calibration: float) -> list[float]:
+               calibration: int,
+               entralink_calibration: int,
+               frame_calibration: int) -> list[float]:
         phases = self.entralink_timer.create(target_delay, target_second, calibration, entralink_calibration)
         phases.append((target_advances / self.ENTRALINK_FRAME_RATE) * 1000 + frame_calibration)
         return phases
